refactor: extract page routes to its own file
