+++ src/pages/pageRoutes.ts
@@ -1,0 +1,10 @@
+export enum PageRoutes {
+  TEAMS = "/teams",
+  TEAM = "/team/:id",
+  JOIN_TEAM = "/join/:joinTokenId",
+  HEARTBEAT = "/heartbeat",
+  CREATE_TEAM = "/create-team",
+  HOME = "/",
+  ERROR = "/error",
+  NOT_FOUND = "/*",
+}
